+++ src/app/COMPONENTS/CustomerSection/products-display/products-display.component.ts
@@ -58,13 +58,19 @@ export class ProductsDisplayComponent implements OnInit {
   }
 
   
- logout(){
-  this.loginService.removeToken();
-  console.log("Log out initiated");
-  this.cartService.removeAllCart();
-  alert('Are ypou sure you want to log out ?');
-  this.router.navigate(['']);
-}
+  logout(){
+    if(this.loginService.isLoggedin()){
+      this.loginService.removeToken();
+      console.log("Log out initiated");
+       this.cartService.removeAllCart();
+      alert('Are you sure you want to log out ?');
+      this.router.navigate(['']);
+    }
+    else{
+      alert("You are not logged in . PLease Login First")
+      this.router.navigate(['/login'])
+    }
+  }
 }
