+++ src/app/COMPONENTS/CustomerSection/view-product-info/view-product-info.component.ts
@@ -46,11 +46,17 @@ export class ViewProductInfoComponent implements OnInit {
     this.cartService.addToCart(product);
   }
   logout(){
-    this.loginService.removeToken();
-    console.log("Log out initiated");
-    this.cartService.removeAllCart();
-    alert('Are you sure you want to log out ?');
-    this.router.navigate(['']);
+    if(this.loginService.isLoggedin()){
+      this.loginService.removeToken();
+      console.log("Log out initiated");
+       this.cartService.removeAllCart();
+      alert('Are you sure you want to log out ?');
+      this.router.navigate(['']);
+    }
+    else{
+      alert("You are not logged in . PLease Login First")
+      this.router.navigate(['/login'])
+    }
   }
